requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# Optional (for some memory operations)
scikit-learn>=1.3.0
